setting views horario
