+++ src/app/app.routes.ts
@@ -40,10 +40,6 @@ export const routes: Routes = [
     path: 'iniciar-estudiante',
     loadComponent: () => import('./pages/estudiante/iniciar/iniciar.component').then((m) => m.IniciarComponent)
   },
-  {
-    path: 'horarios-cursos',
-    loadComponent: () => import('./pages/estudiante/horarios-e/horarios-e.component').then((m) => m.HorariosEComponent)
-  },
   //Route Profesor
   {
     path: 'horario-profesor/:id',
