--- src/app/pages/estudiante/horarios-e/horarios-e.component.css
@@ -1,273 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Rubik:ital,wght@0,300..900;1,300..900&display=swap');
-
-:root{
-  --blue: #C9D7DD;
-  --white: #fff;
-  --light-white: #F1EEDC;
-  --black: #F6F4EB;
-  --light-bg: #164863;
-}
-
-*{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  outline: none;
-  border: none;
-  text-decoration: none;
-  transition: .2s linear;
-  text-transform: capitalize;
-  font-family: 'Rubik', sans-serif;
-}
-
-*::selection{
-  background-color: var(--white);
-  color: var(--black);
-}
-
-html{
-  font-size: 30px;
-  overflow: hidden;
-  scroll-behavior: smooth;
-  scroll-padding-top: 8rem;
-}
-
-::-webkit-scrollbar{
-  height: .5rem;
-  width: 1rem;
-}
-
-::-webkit-scrollbar-track{
-  background-color: transparent;
-}
-
-::-webkit-scrollbar-thumb{
-  background-color: var(--blue);
-  border-radius: 5rem;
-}
-
-body{
-  background-color: var(--black);
-}
-
-section{
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-.header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 1000;
-  background-color: var(--light-bg);
-}
-
-.header .flex{
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.header .flex .logo {
-  font-size: 20px;
-  color: var(--white);
-}
-
-.header .flex .navbar ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  color: var(--light-white);
-  font-size: 20px; 
-}
-
-.header .flex .navbar ul li {
-  margin-left: 2rem;
-}
-
-.header .flex .navbar ul li a {
-  color: var(--light-white);
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.header .flex .navbar ul li a:hover,
-.header .flex .navbar ul li a.active {
-  color: var(--blue);
-}
-
-section {
-    padding: 25px;
-  }
-  
-  .container {
-    margin-top: 1px;
-}
-
-.co {
-    text-align: center;
-    align-items: center;
-    font-size: 18px; 
-    padding-right: 8px;
-    padding-left: 8px;
-    padding-top: 8px;
-    padding-bottom: 8px;
-    background-color: #164863;
-    border: 1px solid #F1EEDC;
-    color: white;
-}
-
-.col {
-    text-align: center;
-    font-size: 14px; 
-    padding-right: 8px;
-    padding-left: 8px;
-    border: 1px solid #F1EEDC; 
-    background: #fff;
-    min-width: 125px; 
-    min-height: 160px; 
-}
-
-table {
-    border-collapse: collapse;
-    width: 100%; 
-    max-width: 200px; 
-    height: auto; 
-    overflow: hidden; 
-    border-radius: 20px; 
-    margin-top: 20px;
-    text-align: center;
-}
-
-.curso-container {
-  display: flex;
-  justify-content: center; 
-  align-items: center; 
-  height: 100%;
-}
-
-.curso-nombre {
-  display: block;
-  text-align: center; 
-}
-
-
-.customBack {
-    position: relative;
-    height: 100vh; 
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    align-items: flex-start; 
-    padding-top: 50px; 
-}
-
-@page {
-    margin: 10mm; 
-}
-
-@media print {
-    @page {
-        size: landscape;
-    }
-}
-
-table tr {
-    line-height: 1; 
-}
-
-
-.out {
-  position: fixed;
-  bottom: 735px;
-  right: 1460px;
-  z-index: 1001;
-  padding: 10px 20px;
-  background-color: #164863; 
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  text-decoration: none; 
-  }
-  
-  .out:hover{
-    background-color: #088395;
-  }
-
-  .info{
-    padding-top: 90px;
-    text-align: center;
-  }
-
-  .info h2{
-    text-align: center;
-    padding-bottom: 40px;
-    font-size: 40px;
-  }
-
-  .profesor-info {
-    display: flex;
-    justify-content: center;
-    gap: 100px; 
-}
-
-.profesor-info p {
-    margin: 0;
-    font-size: 19px;
-}
-
-.outs {
-  cursor: pointer;
-  background-color: #164863; 
-  color: white;
-  border: none;
-  border-radius: 5px;
-  padding: 10px;
-  transform: translateY(-200px) translateX(75px);
-  }
-
-  .curso{
-    padding-top: 75px;
-    align-items: center;
-    text-align: center;
-    display: flex;
-    gap: 20px;
-    padding-left: 70px;
-  }
-
-  .select{
-    font-size: 18px;
-  }
-
-  .info{
-    padding-top: 30px;
-    text-align: center;
-  }
-
-  .info h2{
-    text-align: center;
-    padding-bottom: 40px;
-    font-size: 40px;
-  }
-
-  h2{
-    text-align: center;
-  }
-
-  .info-H {
-    display: flex;
-    justify-content: center;
-    gap: 100px; 
-}
-
-.info-H p {
-    margin: 0;
-    font-size: 19px;
-    color: #000;
-}
