--- src/app/pages/estudiante/horarios-e/horarios-e.component.html
@@ -1,69 +1,0 @@
-<header class="header">
-  <section class="flex">
-    <a href="#Home" class="logo">Centro de Lenguas</a>
-
-    <nav class="navbar">
-      <ul>
-        <li><a routerLink="/horario-estudiante/:id" class="active">Horario</a></li>
-        <li><a routerLink="/horarios-cursos" class="active">Ver Horarios</a></li>
-      </ul>    
-    </nav>
-
-  </section>
-</header>
-
-<section>
-<div class="curso">
-    <h2>Selecciona un curso:</h2>
-    <select class="select" [(ngModel)]="cursoSeleccionado" (change)="mostrarHorario()">
-      <option *ngFor="let curso of cursos" [ngValue]="curso">{{ curso.nombre }}</option>
-    </select>
-  </div>
-</section>
-
-<section>
-  <h2>Horarios</h2>
-  <div class="info">
-  <div class="info-H" *ngIf="cursoSeleccionado">
-    <p>Materia: {{ cursoSeleccionado.nombre }}</p>
-    <p>Profesor: {{ obtenerNombreProfesor(cursoSeleccionado.id_profe) }}</p>
-  </div>
-</div>
-    <div class="customBack">
-      <div  class="container">
-        <div class="row">
-          <button class="out" routerLink="/inicio">Salir</button>
-          <div class="custom-col">
-            <table class="table">
-            <thead>
-                <tr class="row">
-                     <th class="co">Hora</th>
-                     <th class="co">Lunes</th>
-                     <th class="co">Martes</th>
-                     <th class="co">Miércoles</th>
-                     <th class="co">Jueves</th>
-                     <th class="co">Viernes</th>
-                     <th class="co">Sábado</th>
-                     <th class="co">Domingo</th>
-                </tr>
-            </thead>
-            <tbody>
-                <tr *ngFor="let hora of horas">
-                  <td class="col">{{ hora }}</td>
-                    <td class="col" *ngFor="let dia of dias">
-                         <div *ngFor="let item of obtenerHorarioPorDiaYHora(dia, hora)" class="curso-container">
-                             <div class="nombre-y-botones">
-                             <span class="curso-nombre">{{ obtenerNombreCurso(item.curso_id) }}</span>
-                             </div>
-                        </div>                    
-                    </td>
-                </tr>
-            </tbody>
-            </table>
-  </div>
-        </div>
-      </div>
-    </div>
-    </section>
-
-
