+++ src/app/pages/estudiante/iniciar/iniciar.component.html
@@ -69,11 +69,3 @@
         </form>
     </div>
 </div>
-
-
-
-
-
-
-
-
